requirements.txt
pandas
numpy
Pillow
opencv-python
shapely
matplotlib
seaborn